# wgengine/filter: use NewContainsIPFunc for Srcs matches